meta-slack-reporting/requirements.txt
facebook-business==19.0.0
requests==2.31.0
python-dotenv==1.0.0
